pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fox-pro-ai"
version = "4.0.0"
description = "🦊 Fox Pro AI — Professional AI-Native Development Toolkit"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Mickhael", email = "your-email@example.com"}
]
maintainers = [
    {name = "Mickhael"}
]
keywords = [
    "ai", "development", "toolkit", "cursor", "copilot", 
    "claude", "windsurf", "fox", "code-review", "token-optimization"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Code Generators",
    "Topic :: Software Development :: Quality Assurance",
]

dependencies = [
    "pyyaml>=6.0",
    "tiktoken>=0.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.1",
    "pytest-asyncio>=0.23",
    "mypy>=1.8",
    "ruff>=0.3",
    "types-PyYAML>=6.0",
]
web = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.6",
]
gui = [
    # tkinter встроен в Python
]
all = [
    "fox-pro-ai[dev,web]",
]

[project.scripts]
fox = "src.cli:main"
fox-pro-ai = "src.cli:main"

[project.gui-scripts]
fox-gui = "gui.app:run_gui"

[project.urls]
Homepage = "https://github.com/Fox-pro-ai/Fox-pro-ai"
Documentation = "https://github.com/Fox-pro-ai/Fox-pro-ai#readme"
Repository = "https://github.com/Fox-pro-ai/Fox-pro-ai"
Issues = "https://github.com/Fox-pro-ai/Fox-pro-ai/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short --strict-markers"
asyncio_mode = "auto"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
show_error_codes = true
show_column_numbers = true
pretty = true

# Игнорируем пока некоторые ошибки
disallow_untyped_calls = false
disallow_untyped_defs = false
disallow_incomplete_defs = false

[[tool.mypy.overrides]]
module = [
    "yaml",
    "tkinter.*",
]
ignore_missing_imports = true

[tool.ruff]
target-version = "py310"
line-length = 100
exclude = [
    ".git",
    ".venv",
    "venv",
    "__pycache__",
    "build",
    "dist",
    "*.egg-info",
]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "W",   # pycodestyle warnings
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
    "RUF", # ruff-specific
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "N999",   # invalid module name
    "RUF001", # ambiguous unicode character
    "SIM108", # ternary operator
]

[tool.ruff.lint.isort]
known-first-party = ["src", "gui", "plugins"]
force-single-line = false
lines-after-imports = 2

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101", "PLR2004"]
"__init__.py" = ["F401"]

[tool.coverage.run]
source = ["src"]
branch = true
omit = [
    "tests/*",
    "gui/*",
    "*/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]
